rename files following our convention
